feat: update delete to remove associated image
- deleteThing from stuff controller now removes associated image from
filesystem on deletion of a Thing instance.

+++ controllers/stuff.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 const Thing = require('../models/Thing');
 
 exports.createThing = (req, res, next) => {
@@ -51,8 +53,24 @@ exports.modifyThing = (req, res, next) => {
     .catch(error => res.status(400).json({ error }));
 };
 
+// exports.deleteThing = (req, res, next) => {
+//   Thing.deleteOne({ _id: req.params.id})
+//     .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+//     .catch(error => res.status(400).json({ error }));
+// };
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id})
-    .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
+  Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if (thing.userId != req.auth.userId) {
+        res.status(403).json({ error: 'Requête non autorisée' });
+      } else {
+        const filename = thing.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+            .catch(error => res.status(400).json({ error }));
+        });
+      }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
